cluster validation now checks if a relevant exception may apply before failing new image validation
